add performance testing cases to plugin's content dir
